+++ CHANGELOG.md
@@ -65,6 +65,18 @@ Full documentation for amd_smi_lib is available at [https://rocm.docs.amd.com/pr
       - amdsmi_set_power_cap()
   - See the Changed section for changes made to the `set` and `static` commands regarding support for PPT1.  
 
+- **Added PTL Support to `amd-smi static` and `amd-smi set`**
+  - `amd-smi set` now support --ptl-status 0|1 and --ptl-format FORMAT1,FORMAT2
+  - `amd-smi static -l` shows current state and format of PTL
+
+  ```console
+  $ amd-smi static -l
+  GPU: 0
+      LIMIT:
+        PTL_STATE: Enabled
+        PTL_FORMAT: I8,F64
+  ```
+
 ### Changed
 
 - **The `amd-smi` command now shows hsmp rather than amd_hsmp**.